src/pd_fusion/models/calibrate.py
from sklearn.calibration import CalibratedClassifierCV
import numpy as np

def calibrate_model(model, X_val, y_val, method="sigmoid"):
    """
    Apply Platt scaling or Isotonic regression.
    """
    # Placeholder for temperature scaling implementation for PyTorch models
    # Or wrapper using sklearn CalibratedClassifierCV for GBDTs
    print(f"Calibrating model using {method}...")
    return model # Return calibrated model
